+++ packages/chain-adapters/CHANGELOG.md
@@ -1,3 +1,68 @@
+# [@shapeshiftoss/chain-adapters-v2.4.5](https://github.com/shapeshift/lib/compare/@shapeshiftoss/chain-adapters-v2.4.4...@shapeshiftoss/chain-adapters-v2.4.5) (2022-03-17)
+
+# [@shapeshiftoss/chain-adapters-v2.4.4](https://github.com/shapeshift/lib/compare/@shapeshiftoss/chain-adapters-v2.4.3...@shapeshiftoss/chain-adapters-v2.4.4) (2022-03-16)
+
+
+### Bug Fixes
+
+* update unchained client ([#451](https://github.com/shapeshift/lib/issues/451)) ([d23e179](https://github.com/shapeshift/lib/commit/d23e17995889343addf6bbcb79744ad11e3f8127))
+
+# [@shapeshiftoss/chain-adapters-v2.4.3](https://github.com/shapeshift/lib/compare/@shapeshiftoss/chain-adapters-v2.4.2...@shapeshiftoss/chain-adapters-v2.4.3) (2022-03-16)
+
+
+### Bug Fixes
+
+* btc getAddress uses the wrong accountType ([#447](https://github.com/shapeshift/lib/issues/447)) ([108851d](https://github.com/shapeshift/lib/commit/108851d015e17a351b6d04b6c42544f363f4f591))
+
+# [@shapeshiftoss/chain-adapters-v2.4.2](https://github.com/shapeshift/lib/compare/@shapeshiftoss/chain-adapters-v2.4.1...@shapeshiftoss/chain-adapters-v2.4.2) (2022-03-16)
+
+
+### Bug Fixes
+
+* update json rpc ([#442](https://github.com/shapeshift/lib/issues/442)) ([abfb16c](https://github.com/shapeshift/lib/commit/abfb16c3d28fa40bbb20e9834d95dca9d13e008a))
+
+# [@shapeshiftoss/chain-adapters-v2.4.1](https://github.com/shapeshift/lib/compare/@shapeshiftoss/chain-adapters-v2.4.0...@shapeshiftoss/chain-adapters-v2.4.1) (2022-03-15)
+
+
+### Bug Fixes
+
+* removed error from required chainId and set on class ([#426](https://github.com/shapeshift/lib/issues/426)) ([8693d21](https://github.com/shapeshift/lib/commit/8693d217ed75f3f99882e8174653d5b95c9cfe97))
+
+# [@shapeshiftoss/chain-adapters-v2.4.0](https://github.com/shapeshift/lib/compare/@shapeshiftoss/chain-adapters-v2.3.0...@shapeshiftoss/chain-adapters-v2.4.0) (2022-03-14)
+
+
+### Features
+
+* cosmos unsubscribeTxs and closeTxs ([#434](https://github.com/shapeshift/lib/issues/434)) ([95e4ee1](https://github.com/shapeshift/lib/commit/95e4ee1ffc92a1d66348435937cef0401796d2d9))
+
+# [@shapeshiftoss/chain-adapters-v2.3.0](https://github.com/shapeshift/lib/compare/@shapeshiftoss/chain-adapters-v2.2.0...@shapeshiftoss/chain-adapters-v2.3.0) (2022-03-10)
+
+
+### Bug Fixes
+
+* update argument type ([#428](https://github.com/shapeshift/lib/issues/428)) ([08feb5e](https://github.com/shapeshift/lib/commit/08feb5e4e80c5329f4a34d84715086eea6beb722))
+
+
+### Features
+
+* cosmos tx history support ([#416](https://github.com/shapeshift/lib/issues/416)) ([94f24eb](https://github.com/shapeshift/lib/commit/94f24ebfe83f36bf0802d99bdcf7626905b82432))
+
+# [@shapeshiftoss/chain-adapters-v2.2.0](https://github.com/shapeshift/lib/compare/@shapeshiftoss/chain-adapters-v2.1.0...@shapeshiftoss/chain-adapters-v2.2.0) (2022-03-08)
+
+
+### Features
+
+* utxo chain adapter refactor ([#411](https://github.com/shapeshift/lib/issues/411)) ([c6d1b31](https://github.com/shapeshift/lib/commit/c6d1b311eefa3b7f48e24ef763b41eab27045191))
+
+# [@shapeshiftoss/chain-adapters-v2.1.0](https://github.com/shapeshift/lib/compare/@shapeshiftoss/chain-adapters-v2.0.2...@shapeshiftoss/chain-adapters-v2.1.0) (2022-03-08)
+
+
+### Features
+
+* cosmos `subscribeTxs` ([#415](https://github.com/shapeshift/lib/issues/415)) ([77fd091](https://github.com/shapeshift/lib/commit/77fd091a3f8a7927d90956c466c6dad002216db9))
+
+# [@shapeshiftoss/chain-adapters-v2.0.2](https://github.com/shapeshift/lib/compare/@shapeshiftoss/chain-adapters-v2.0.1...@shapeshiftoss/chain-adapters-v2.0.2) (2022-03-04)
+
 # [@shapeshiftoss/chain-adapters-v2.0.1](https://github.com/shapeshift/lib/compare/@shapeshiftoss/chain-adapters-v2.0.0...@shapeshiftoss/chain-adapters-v2.0.1) (2022-03-03)
 
 # [@shapeshiftoss/chain-adapters-v2.0.0](https://github.com/shapeshift/lib/compare/@shapeshiftoss/chain-adapters-v1.36.0...@shapeshiftoss/chain-adapters-v2.0.0) (2022-03-03)